+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adapter/AnnotationHandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.techcourse.support.web.adapter;
+package webmvc.org.springframework.web.servlet.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
