+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/view/JsonView.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.view;
+package webmvc.org.springframework.web.servlet.mvc.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
