+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/adapter/HandlerAdaptersTest.java
@@ -1,16 +1,15 @@
-package com.techcourse.support.web.adapter;
+package webmvc.org.springframework.web.servlet.mvc.adapter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.techcourse.controller.LoginController;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
+import samples.TestLegacyController;
 import webmvc.org.springframework.web.servlet.mvc.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.mvc.adapter.HandlerAdapters;
 import webmvc.org.springframework.web.servlet.mvc.view.resolver.ViewResolvers;
 
 @SuppressWarnings("NonAsciiCharacters")
@@ -19,7 +18,7 @@ class HandlerAdaptersTest {
 
     @Test
     void 생성자는_호출하면_HandlerAdapters를_초기화한다() {
-        assertDoesNotThrow(HandlerAdapters::new);
+        Assertions.assertDoesNotThrow(HandlerAdapters::new);
     }
 
     @Test
@@ -38,7 +37,7 @@ class HandlerAdaptersTest {
         viewResolvers.initialize();
         handlerAdapters.initialize(viewResolvers);
 
-        final HandlerAdapter actual = handlerAdapters.getHandlerAdapter(new LoginController());
+        final HandlerAdapter actual = handlerAdapters.getHandlerAdapter(new TestLegacyController());
 
         assertThat(actual).isNotNull();
     }
@@ -47,7 +46,7 @@ class HandlerAdaptersTest {
     void getHandlerAdapter_메서드는_해당_핸들러를_처리할_수_있는_HandlerAdapter가_없으면_예외가_발생한다() {
         final HandlerAdapters handlerAdapters = new HandlerAdapters();
 
-        assertThatThrownBy(() -> handlerAdapters.getHandlerAdapter(new LoginController()))
+        assertThatThrownBy(() -> handlerAdapters.getHandlerAdapter(new TestLegacyController()))
                 .isInstanceOf(IllegalStateException.class)
                 .hasMessageContaining("해당 Handler를 수행할 수 있는 HandlerAdapter가 존재하지 않습니다.");
     }
